Extract `Crate` to `test_helpers`
